修复ci错误

See merge request SilverBlog/SilverBlog!182

+++ manage/theme.py
@@ -48,19 +48,25 @@ def install_theme(theme_name, custom):
         git_repo = theme_name
     try:
         git.Repo.clone_from(url=git_repo, to_path=repo_dir, depth=1)
-    except git.exc.GitCommandError as err:
+    except git.exc.GitCommandError:
         console.log("Error", "Unable to clone theme repository.")
         exit(1)
-    config_example_file = "{}/config.example.json".format(repo_dir)
-    static_symlink = "./templates/static/{}".format(theme_name)
+    if os.path.exists("{}/package-metadata.json".format(repo_dir)):
+        config_example_file = "{}/config.example.json".format(repo_dir)
+        static_symlink = "./templates/static/{}".format(theme_name)
 
-    if not os.path.exists(static_symlink):
-        os.symlink("{}/templates/{}/static".format(os.getcwd(), theme_name), static_symlink)
+        if not os.path.exists(static_symlink):
+            os.symlink("{}/templates/{}/static".format(os.getcwd(), theme_name), static_symlink)
 
-    if os.path.exists(config_example_file):
-        shutil.copyfile(config_example_file, "{}/config.json".format(repo_dir))
+        if os.path.exists(config_example_file):
+            shutil.copyfile(config_example_file, "{}/config.json".format(repo_dir))
 
-    download_static_file(theme_name)
+        download_static_file(theme_name)
+    if not os.path.exists("{}/package-metadata.json".format(repo_dir)) and os.path.exists("{}/install.sh".format(repo_dir)):
+        result_code = os.system("cd templates && bash {}/install.sh".format(theme_name))
+        if (result_code >> 8) != 0:
+            console.log("Error", "An error occurred while executing the install script.")
+            exit(1)
     console.log("Success", "The theme is install successfully!")
 
 
@@ -82,14 +88,17 @@ def upgrade_theme(theme_name):
         if repo.rev_parse("HEAD") != repo.rev_parse("FETCH_HEAD"):
             console.log("Info", "Updating theme, please wait...")
             remote.pull()
-            download_static_file(theme_name)
+            if os.path.exists("./templates/{}/package-metadata.json".format(theme_name)):
+                download_static_file(theme_name)
             console.log("Success", "The theme is upgrade successfully!")
             return
     console.log("Info", "No upgrade found.")
 
 
 def download_static_file(theme_name):
-    download_list_file = "./templates/{}/package.json".format(theme_name)
+    if not os.path.exists("./templates/{}/package-metadata.json".format(theme_name)):
+        return
+    download_list_file = "./templates/{}/package-metadata.json".format(theme_name)
     download_file_location = "./templates/{}/static/library".format(theme_name)
     if os.path.exists(download_file_location):
         shutil.rmtree(download_file_location)
